+++ release-notes.md
@@ -3,9 +3,10 @@ RELEASE NOTES
 **Release Notes for Testy 2.1.0-SNAPSHOT**
 - use attributes when generate toString, in case no other important string is present
 - add support for css selector when setExcludeClasses
-- update webdriver version 2.49.0
+- update webdriver version 2.52.0
 - add log at sendKeys method
 - add possibility to overwritten value from webLocator.properties in browser.properties
+- change getHtmlText() in getText()
 
 **Release Notes for Testy 2.0.0**
 - Change config default value for weblocator.log.suggestions=false